magic wand added
