+++ apps/frontend/app/(website)/studio/features/modGPT/PromptPanel/GenerateTestCasesButton.tsx
@@ -1,5 +1,7 @@
+import { cn } from "@/utils";
 import type { useCodemodAI } from "@chatbot/useAiService/codemodAI/useCodemodAI";
-import ButtonWithTooltip from "@studio/components/button/BottonWithTooltip";
+import { MagicWand } from "@phosphor-icons/react";
+import Tooltip from "@studio/components/Tooltip/Tooltip";
 
 export const GenerateTestCasesButton = ({
   handleButtonClick,
@@ -7,19 +9,24 @@ export const GenerateTestCasesButton = ({
   handleButtonClick: ReturnType<typeof useCodemodAI>["autogenerateTestCases"];
 }) => {
   return (
-    <ButtonWithTooltip
-      tooltipContent={
-        <>
+    <Tooltip
+      trigger={
+        <button
+          onClick={handleButtonClick}
+          className={cn(
+            "cursor-pointer border-hidden align-text-top bg-transparent hover:bg-transparent p-3",
+          )}
+        >
+          <MagicWand size={"30px"} />
+        </button>
+      }
+      content={
+        <p>
+          {" "}
           Generate a new pair of before/after based on your existing code
           examples OR based on the natural language description.
-        </>
+        </p>
       }
-      variant="default"
-      size="sm"
-      className="bg-white text-black flex gap-1 text-xs my-0 h-8 !py-0 hover:bg-accent hover:text-black"
-      onClick={handleButtonClick}
-    >
-      Autogenerate test cases
-    </ButtonWithTooltip>
+    />
   );
 };
