+++ apps/frontend/app/(website)/studio/main/5PaneSetup.tsx
@@ -1,5 +1,6 @@
 import { LEARN_KEY } from "@/constants";
 import { cn } from "@/utils";
+import { useAiService } from "@chatbot/useAiService";
 import type { KnownEngines } from "@codemod-com/utilities";
 import { useTheme } from "@context/useTheme";
 import { getCodeDiff } from "@studio/api/getCodeDiff";
@@ -19,7 +20,9 @@ import { TestTabsComponent } from "@studio/main/PageBottomPane/TestTabsComponent
 import { AssistantTab } from "@studio/main/PaneLayout";
 import { LoginWarningModal } from "@studio/main/PaneLayout/LoginWarningModal";
 import { enginesConfig } from "@studio/main/PaneLayout/enginesConfig";
+import { useCFSStore } from "@studio/store/CFS";
 import { SEARCH_PARAMS_KEYS } from "@studio/store/initialState";
+import { useModStore } from "@studio/store/mod";
 import { useSnippetsStore } from "@studio/store/snippets";
 import { useEffect, useRef } from "react";
 import { PanelGroup } from "react-resizable-panels";
@@ -40,7 +43,15 @@ const Main = () => {
 
   const { engine, setEngine, getSelectedEditors } = useSnippetsStore();
   const { isDark } = useTheme();
+  const { setContent } = useModStore();
+  const {
+    AIAssistant: { engine: llmEngine },
+  } = useCFSStore();
 
+  const aiAssistantData = useAiService({
+    setCodemod: setContent,
+    engine: llmEngine,
+  });
   const onEngineChange = (value: (typeof enginesConfig)[number]["value"]) => {
     setEngine(value as KnownEngines);
   };
@@ -167,6 +178,7 @@ const Main = () => {
       className="bg-gray-bg assistant"
     >
       <AssistantTab
+        aiAssistantData={aiAssistantData}
         panelRefs={panelRefs}
         beforePanel={beforePanel}
         afterPanel={afterPanel}
@@ -232,7 +244,9 @@ const Main = () => {
               panelRefs={panelRefs}
               className="bg-gray-bg"
             >
-              <TestTabsComponent />
+              <TestTabsComponent
+                autogenerateTestCases={aiAssistantData.autogenerateTestCases}
+              />
               <PanelGroup direction="horizontal">
                 {beforeAndAfterPanel}
                 <ResizeHandle direction="horizontal" />
